requirements.txt
asgiref==3.8.1
boto3==1.38.25
botocore==1.38.25
colorama==0.4.6
coverage==7.7.1
dj-database-url==2.1.0
Django==5.1.7
django-cors-headers==4.7.0
django-rest-passwordreset==1.5.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
git-filter-repo==2.47.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
openpyxl==3.1.5
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
xlwt==1.3.0
